delete_task function complete

+++ FloorPlan/views.py
@@ -74,7 +74,8 @@ def edit_task(request, pk):
 
 def delete_task(request, pk):
     task = get_object_or_404(Task, pk=pk)
-    task.delete()
+    if request.method == 'POST':
+        task.delete()
     return redirect('task-detail', pk=task.project.pk)
 
 def new_team_member(request, pk):
